Updated JS code to ES6 and added date
