+++ modules/creat_book.js
@@ -1,0 +1,6 @@
+export class StandardBook {
+    constructor(title, author) {
+      this.title = title;
+      this.author = author;
+    }
+}
